+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Changed
 - `JsonMachine\JsonMachine` entry point class is deprecated, use `JsonMachine\Items` instead.
-- `JsonMachine\JsonDecoder\Decoder` interface is deprecated. Use `JsonMachine\JsonDecoder\ChunkDecoder` instead. 
+- `JsonMachine\JsonDecoder\Decoder` interface is deprecated. Use `JsonMachine\JsonDecoder\ChunkDecoder` instead.
+- Internal decoders moved to `ChunkDecoder`. `ErrorWrappingDecoder` decorator now requires `ChunkDecoder` as well. 
 
 ### Added
 - New entry point class `Items` replaces `JsonMachine`.